addblog and editblog are same
