+++ new_client/src/App.js
@@ -14,12 +14,11 @@ import Event from "./components/events/Event";
 
 import Blogs from "./components/blogs/Blog";
 import AddBlog from "./components/blogs/AddBlog";
+import IndividualBlog from "./components/blogs/IndividualBlog";
 
 import Signin from "./components/auth/Signin.jsx";
 import Signup from "./components/auth/Signup.jsx";
 import NewHome from "./components/Home/Newhome";
-import IndividualBlog from "./components/blogs/IndividualBlog";
-import EditBlog from "./components/blogs/EditBlog";
 
 function App() {
   const store = configureStore();
@@ -37,7 +36,7 @@ function App() {
             <Route path="/signin" component={Signin} />
             <Route exact path="/addblog" component={AddBlog} />
             <Route path="/blogs/:id" component={IndividualBlog} />
-            <Route path="/blog/edit/:id" component={EditBlog} />
+            <Route path="/blog/edit/:id" component={AddBlog} />
             <Route path="/signup" component={Signup} />
             <Route path="/glimpse" component={Glimpse} />
             <Route path="/newHome" component={NewHome} />
