+++ new_client/src/components/blogs/AddBlog.jsx
@@ -1,7 +1,7 @@
 import "date-fns";
 import DateFnsUtils from "@date-io/date-fns";
 import { connect } from "react-redux";
-import React, { useState } from "react";
+import React, { useState, useEffect } from "react";
 import Container from "@material-ui/core/Container";
 import { TextField, Button, Grid } from "@material-ui/core";
 import {
@@ -11,8 +11,11 @@ import {
 import Editor from "./Editor";
 import axios from "axios";
 import AlertUtility from "../Utilities/Alert";
+import { useParams } from "react-router-dom";
 
 function AddBlog(props) {
+  const id = useParams().id;
+
   const [blogTitle, setBlogTitle] = useState("");
   const [blogAuthor, setBlogAuthor] = useState("");
   const [blogContent, setBlogContent] = useState("**Hello world!!!**");
@@ -20,6 +23,12 @@ function AddBlog(props) {
 
   const [isError, setIsError] = useState(false);
   const [isSubmitted, setIsSubmitted] = useState(false);
+
+  const isEditPage = !!id;
+
+  const successString = isEditPage
+    ? "Blog edited successfully!"
+    : "Blog added successfully!";
 
   const handleDateChange = (date) => {
     setSelectedDate(date);
@@ -42,7 +51,7 @@ function AddBlog(props) {
       .post(process.env.REACT_APP_API + "/blogs/new", JSON.stringify(blog), {
         headers: {
           "Content-Type": "application/json",
-          "Authorization": "Bearer " + props.token
+          Authorization: "Bearer " + props.token,
         },
       })
       .then((res) => {
@@ -55,6 +64,41 @@ function AddBlog(props) {
       .catch((err) => setIsError(true));
   };
 
+  const handleDataEdit = async () => {
+    const blog = {
+      blogTitle: blogTitle,
+      blogContent: blogContent,
+    };
+    const res = await axios.put(
+      process.env.REACT_APP_API + `/blogs/edit/${id}`,
+      JSON.stringify(blog),
+      {
+        headers: {
+          "Content-Type": "application/json",
+          Authorization: "Bearer " + props.token,
+        },
+      }
+    );
+    if (res.status !== 200) {
+      setIsError(true);
+    } else {
+      setIsSubmitted(true);
+    }
+  };
+
+  useEffect(() => {
+    axios
+      .get(process.env.REACT_APP_API + `/blogs/${id}`)
+      .then((res) => {
+        const { blogTitle, blogContent, date, author } = res.data;
+        setBlogTitle(blogTitle);
+        setBlogContent(blogContent);
+        setBlogAuthor(author);
+        setSelectedDate(date);
+      })
+      .catch((err) => console.log(err.toString()));
+  }, [id]);
+
   return (
     <Container maxWidth="md" style={{ backgroundColor: "white" }}>
       <Grid container spacing={3}>
@@ -66,6 +110,7 @@ function AddBlog(props) {
               label="Date picker dialog"
               format="MM/dd/yyyy"
               value={selectedDate}
+              disabled={isEditPage}
               onChange={handleDateChange}
               KeyboardButtonProps={{
                 "aria-label": "change date",
@@ -78,6 +123,7 @@ function AddBlog(props) {
             autoFocus
             label="Enter your username"
             value={blogAuthor}
+            disabled={isEditPage}
             onChange={(e) => setBlogAuthor(e.target.value)}
           />
         </Grid>
@@ -95,20 +141,20 @@ function AddBlog(props) {
           <Button
             variant="contained"
             color="primary"
-            onClick={handleDataSubmit}
+            onClick={isEditPage ? handleDataEdit : handleDataSubmit}
           >
             Submit
           </Button>
         </Grid>
       </Grid>
-      <AlertUtility 
+      <AlertUtility
         open={isSubmitted}
         duration={3000}
         onCloseHandler={handleClose}
         severity="success"
-        message="Blog added successfully! Redirecting to Blogs page..."
+        message={successString + " Reloading chats..."}
       />
-      <AlertUtility 
+      <AlertUtility
         open={isError}
         duration={1000}
         onCloseHandler={() => setIsError(false)}
@@ -121,7 +167,7 @@ function AddBlog(props) {
 
 const mapStateToProps = (state) => ({
   userID: state.auth.userID,
-  token: state.auth.token
+  token: state.auth.token,
 });
 
 export default connect(mapStateToProps)(AddBlog);
