--- new_client/src/components/blogs/EditBlog.jsx
@@ -1,140 +1,0 @@
-import "date-fns";
-import DateFnsUtils from "@date-io/date-fns";
-import {
-  MuiPickersUtilsProvider,
-  KeyboardDatePicker,
-} from "@material-ui/pickers";
-import React, { useState, useEffect } from "react";
-import {connect} from "react-redux"
-import { useParams, useHistory } from "react-router-dom";
-import axios from "axios";
-import {
-  Container,
-  Grid,
-  Button,
-  TextField,
-} from "@material-ui/core";
-import Editor from "./Editor";
-import AlertUtility from "../Utilities/Alert";
-
-function EditBlog(props) {
-  const id = useParams().id;
-  const history = useHistory();
-
-  const [blogTitle, setBlogTitle] = useState("");
-  const [blogAuthor, setBlogAuthor] = useState("");
-  const [blogContent, setBlogContent] = useState("");
-  const [selectedDate, setSelectedDate] = useState();
-  const [isSubmitted, setIsSubmitted] = useState(false);
-  const [isError, setIsError] = useState(false);
-
-  useEffect(() => {
-    axios
-      .get(process.env.REACT_APP_API + `/blogs/${id}`)
-      .then((res) => {
-        const { blogTitle, blogContent, date, author } = res.data;
-        setBlogTitle(blogTitle);
-        setBlogContent(blogContent);
-        setBlogAuthor(author);
-        setSelectedDate(date);
-      })
-      .catch((err) => console.log(err.toString()));
-  }, [id]);
-
-  const handleDataSubmit = async () => {
-    const blog = {
-      blogTitle: blogTitle,
-      blogContent: blogContent,
-    };
-    const res = await axios.put(
-      process.env.REACT_APP_API + `/blogs/edit/${id}`,
-      JSON.stringify(blog),
-      {
-        headers: {
-          "Content-Type": "application/json",
-          "Authorization": "Bearer " + props.token
-        },
-      }
-    );
-    if (res.status !== 200) {
-      setIsError(true);
-    } else {
-      setIsSubmitted(true);
-    }
-  };
-
-  const handleClose = () => {
-    setIsSubmitted(false);
-    history.push("/blogs");
-  };
-
-  return (
-    <Container maxWidth="md" style={{ backgroundColor: "white" }}>
-      <Grid container spacing={3}>
-        <Grid item xs={12}>
-          <MuiPickersUtilsProvider utils={DateFnsUtils}>
-            <KeyboardDatePicker
-              margin="normal"
-              id="date-picker-dialog"
-              label="Date posted"
-              format="MM/dd/yyyy"
-              value={selectedDate}
-              disabled
-              KeyboardButtonProps={{
-                "aria-label": "change date",
-              }}
-            />
-          </MuiPickersUtilsProvider>
-        </Grid>
-        <Grid item xs={12}>
-          <TextField
-            autoFocus
-            disabled
-            label="Your username"
-            value={blogAuthor}
-          />
-        </Grid>
-        <Grid item xs={12}>
-          <TextField
-            label="Edit title"
-            value={blogTitle}
-            onChange={(e) => setBlogTitle(e.target.value)}
-          />
-        </Grid>
-        <Grid item xs={12}>
-          <Editor content={blogContent} setContent={setBlogContent} />
-        </Grid>
-        <Grid item xs={12}>
-          <Button
-            variant="contained"
-            color="primary"
-            onClick={handleDataSubmit}
-          >
-            Submit
-          </Button>
-        </Grid>
-      </Grid>
-      <AlertUtility 
-        open={isSubmitted}
-        duration={3000}
-        onCloseHandler={handleClose}
-        severity="success"
-        message="Edited Successfully! Redirecting to Blogs page..."
-      />
-      <AlertUtility 
-        open={isError}
-        duration={1000}
-        onCloseHandler={() => setIsError(false)}
-        severity="error"
-        message="Oops! An error occurred. Please try again."
-      />
-    </Container>
-  );
-}
-
-const mapStateToProps = (state) => ({
-  token: state.auth.token,
-  userID: state.auth.userID
-})
-
-export default connect(mapStateToProps)(EditBlog);
